Remove all files related to the old message bus infrastructure
Removed all files and code related to the old message bus infrastructure to avoid confusion and redundancy. This specifically includes classes and interfaces related to commands, events, consumers, observers and service setup among others. This cleanup is part of the effort to transition to a new event bus infrastructure.

+++ src/Services/Account/Command/Domain/Abstractions/Aggregates/AggregateRoot.cs
@@ -10,7 +10,7 @@ public abstract class AggregateRoot<TValidator> : Entity<TValidator>, IAggregate
 {
     private readonly List<IDomainEvent> _events = new();
 
-    public long Version { get; private set; }
+    public uint Version { get; private set; }
 
     [JsonIgnore]
     public IEnumerable<IDomainEvent> UncommittedEvents
@@ -27,10 +27,10 @@ public abstract class AggregateRoot<TValidator> : Entity<TValidator>, IAggregate
 
     public abstract void Handle(ICommand command);
 
-    protected void RaiseEvent<TEvent>(Func<long, TEvent> func) where TEvent : IDomainEvent
-        => RaiseEvent((func as Func<long, IDomainEvent>)!);
+    protected void RaiseEvent<TEvent>(Func<uint, TEvent> func) where TEvent : IDomainEvent
+        => RaiseEvent((func as Func<uint, IDomainEvent>)!);
 
-    protected void RaiseEvent(Func<long, IDomainEvent> onRaise)
+    protected void RaiseEvent(Func<uint, IDomainEvent>onRaise)
     {
         Version++;
         var @event = onRaise(Version);
